feature: 사용쟈 confirm 받고 삭제 진행

+++ src/pages/payments/cards/list/Cards.tsx
@@ -49,6 +49,9 @@ export const Cards = () => {
   };
 
   const handleDeleteClick = async (id: string) => {
+    const confirm = window.confirm('정말 삭제하시겠습니까?');
+    if (!confirm) return;
+
     await deleteCards({ id });
     refetch();
   };
